refactor(backend): move OS socket probes from model/ports.py to util.py (#1547)
port_in_use, free_port, scan_free_ports, and MAX_PORT are pure OS-level
socket probes that never touch the database; they were misplaced in the
model persistence layer alongside the port_allocations table queries.
Move them to klangk_backend/util.py, which already owns the loopback /
network-trust helpers they are adjacent to.

model/ports.py keeps the DB-backed functions plus find_and_allocate_ports
(the composer that bridges model + util) and re-imports the moved symbols
via __all__, so all historical import paths keep working:
  - from klangk_backend import model; model.port_in_use(...)
  - from klangk_backend.model.ports import port_in_use
  - from klangk_backend.model import free_port   (20+ e2e files)

port_in_use keeps its deliberate 0.0.0.0 bind: workspace host ports are
published by podman with no host IP (i.e. on 0.0.0.0), so the probe must
detect a bind on any interface to predict a publish collision. The
rationale is now captured in the function docstring — CodeQL alert #155
(py/bind-socket-all-network-interfaces) is a false positive.

Tests: the 5 monkeypatches in test_model.py retarget model.ports ->
util (where scan_free_ports now resolves port_in_use), and the direct
port-probe tests move to test_util.py as TestPortDiscovery.

Refs #1547.
