+++ src/backend/klangk_backend/util.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 import asyncio
 import ipaddress
 import logging
+import socket
 import subprocess
 from typing import TypeVar
 
@@ -111,6 +112,84 @@ def sanitize_disposition_name(name: str) -> str:
     (double quotes, backslashes, path separators).
     """
     return name.replace("/", "_").replace("\\", "_").replace('"', "")
+
+
+# --- OS-level TCP port discovery (moved from model/ports.py, #1547) -----
+# These are pure socket probes — they never touch the DB, so they live in
+# util (next to the loopback / network-trust helpers) rather than in the
+# ``model`` persistence layer. ``model.ports`` re-exports them for back-compat.
+
+# Highest valid TCP port.  scan_free_ports will not scan past this, so an
+# exhausted range fails fast instead of looping forever.
+MAX_PORT = 65535
+
+
+def port_in_use(port: int) -> bool:
+    """Check if a port is bound at the OS level.
+
+    Binds ``0.0.0.0`` (all interfaces) deliberately: workspace host ports
+    are published by podman with no host IP, i.e. on ``0.0.0.0`` (see
+    ``podman.create``'s ``-p host:container``), so the probe must detect a
+    bind on *any* interface to predict a publish collision. Binding the
+    probe to loopback would miss ports held only on an external interface
+    and let the allocator hand out a port podman then fails to bind. This
+    is why CodeQL's ``py/bind-socket-all-network-interfaces`` (alert #155)
+    is a false positive here.
+    """
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        try:
+            s.bind(("0.0.0.0", port))
+            return False
+        except OSError:
+            return True
+
+
+def free_port() -> int:
+    """Return a free TCP port on loopback for ephemeral use.
+
+    Binds ``127.0.0.1:0`` so the OS assigns an ephemeral port, then
+    releases it and returns the number. Used by the E2E harnesses to
+    pick the server port (and to seed ``KLANGK_PORT_RANGE_START``)
+    instead of a hardcoded value, so concurrent runs — xdist workers,
+    or several suites on one machine — don't collide (#1393). This
+    generalizes the ``_find_free_port`` helper first introduced in
+    ``test_nginx_acl_e2e.py``.
+
+    The port is released before this returns, so there is an inherent
+    TOCTOU window before the caller rebinds it (e.g. uvicorn at server
+    startup, or a workspace container binding a hosted-app port). For
+    the workspace-port range the allocator's own :func:`port_in_use`
+    check (run inside :func:`scan_free_ports`) is the backstop: it skips
+    any port a concurrent run grabbed in the meantime.
+    """
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        # Loopback (not INADDR_ANY "") for ephemeral pickup — same
+        # free-port behavior, matches the test_nginx_acl_e2e pattern.
+        s.bind(("127.0.0.1", 0))
+        return s.getsockname()[1]
+
+
+def scan_free_ports(start: int, count: int, used: set[int]) -> list[int]:
+    """Find ``count`` free ports at or after ``start``.
+
+    Skips ports already in ``used`` (DB-allocated) and ports reported as
+    bound by the OS.  This is synchronous because it performs blocking
+    ``socket.bind()`` checks; ``model.find_and_allocate_ports`` runs it in
+    an executor so the event loop is not stalled.  Raises ``ValueError`` if
+    fewer than ``count`` free ports are available before ``MAX_PORT``.
+    """
+    ports: list[int] = []
+    port = start
+    while len(ports) < count:
+        if port > MAX_PORT:
+            raise ValueError(
+                f"Could not allocate {count} free ports starting at "
+                f"{start}: exhausted at {MAX_PORT}"
+            )
+        if port not in used and not port_in_use(port):
+            ports.append(port)
+        port += 1
+    return ports
 
 
 # Loopback addresses used by ``Util.client_is_loopback`` (the none-mode
